Move webauthn model into model package.
