+++ core/src/commonMain/kotlin/com/ustadmobile/core/domain/credentials/passkey/model/AuthenticationResponseJSON.kt
@@ -1,4 +1,4 @@
-package com.ustadmobile.core.domain.credentials.passkey.webAuthn
+package com.ustadmobile.core.domain.credentials.passkey.model
 
 import kotlinx.serialization.Serializable
 //during registration and signin with passkey the standard webAuthN response
